14052025 be

+++ gpos4backend/customermgmt/models.py
@@ -9,8 +9,8 @@ class CustomerMaster(models.Model):
     customer_master_code = models.IntegerField(blank=False, null=False) # 1,2,3,4 [Coz CustomerMaster_id will not be serially assigned to every different business]
     customer_name = models.CharField(null=False, blank=False, max_length=120)
     customer_account_id = models.ForeignKey(AccountsMaster, on_delete=models.CASCADE,  blank=False, null=False)
-    customer_phone = models.IntegerField(null=True, blank=True)
-    customer_whatsapp = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True) # Set default to 0
+    customer_phone = models.BigIntegerField(null=True, blank=True)
+    customer_whatsapp = models.BigIntegerField(null=True, blank=True) # Set default to 0
     customer_add1 = models.CharField(max_length=220, null=True, blank=True)
     customer_add2 = models.CharField(max_length=30, null=True, blank=True)
     customer_city = models.CharField(max_length=30, null=True, blank=True)
